[#issue 25] carga de Fixtures
añado bundles de fixtures a app/autoload.php y quito dos archivos que están en gitignore

añado bundles de fixtures a app/AppKernel.php
